+++ .github/workflows/integration-tests.yml
@@ -25,6 +25,7 @@ jobs:
           [
             github.com/ChainSafe/gossamer/dot,
             github.com/ChainSafe/gossamer/dot/core,
+            github.com/ChainSafe/gossamer/dot/rpc,
             github.com/ChainSafe/gossamer/dot/rpc/modules,
             github.com/ChainSafe/gossamer/lib/babe,
             github.com/ChainSafe/gossamer/dot/sync,
