+++ dot/rpc/websocket_integration_test.go
@@ -1,3 +1,6 @@
+//go:build integration
+// +build integration
+
 // Copyright 2021 ChainSafe Systems (ON)
 // SPDX-License-Identifier: LGPL-3.0-only
